requirements.txt
asgiref==3.8.1
attrs==23.2.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.5
defusedxml==0.7.1
deprecation==2.1.0
Django==5.0.4
django-allauth==0.61.1
django-email-verification==0.3.3
django-use-email-as-username==1.4.0
djangorestframework==3.15.1
idna==3.6
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
oauthlib==3.2.2
packaging==24.0
pillow==10.3.0
pycparser==2.22
PyJWT==2.8.0
pyproject-toml==0.0.10
python-dotenv==1.0.1
python3-openid==3.2.0
referencing==0.34.0
requests==2.31.0
requests-oauthlib==2.0.0
rpds-py==0.18.0
setuptools==69.2.0
sqlparse==0.4.4
toml==0.10.2
tzdata==2024.1
urllib3==2.2.1
validators==0.27.0
wheel==0.43.0
